% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cdcfluview-package.R
\docType{package}
\name{cdcfluview}
\alias{cdcfluview}
\alias{cdcfluview-package}
\title{A package to retrive data behind the CDC FluView portal}
\description{
A package to retrive data behind the CDC FluView portal
}
\author{
Bob Rudis (@hrbrmstr)
}

