% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{census_regions}
\alias{census_regions}
\title{Census Region Table}
\format{
A data frame with 51 rows and 2 variables
}
\usage{
data(census_regions)
}
\description{
This dataset contains the states belonging to the (presently) 4
             U.S. Census regions in "long" format. It consists of a \code{data.frame}
             with the following columns:

\itemize{
  \item \code{region}: the official Census region name (e.g. "\code{East}")
  \item \code{state}: state belonging to the region
}
}
\note{
Last updated 2015-08-09.
}
\references{
\url{https://www.cdc.gov/std/stats12/images/CensusMap.png}
}
\keyword{datasets}
