% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographic-spread.R
\name{geographic_spread}
\alias{geographic_spread}
\title{State and Territorial Epidemiologists Reports of Geographic Spread of Influenza}
\usage{
geographic_spread()
}
\description{
State and Territorial Epidemiologists Reports of Geographic Spread of Influenza
}
\examples{
\dontrun{
geographic_spread()
}
}
