% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{hhs_regions}
\alias{hhs_regions}
\title{HHS Region Table}
\format{A data frame with 59 rows and 4 variables}
\usage{
data(hhs_regions)
}
\description{
This dataset contains the names, numbers, regional offices for-,
             and states/territories belonging to the (presently) 10 HHS U.S.
             regions in "long" format. It consists of a \code{data.frame}
             with the following columns:

\itemize{
  \item \code{region}: the official HHS region name (e.g. "\code{Region 1}")
  \item \code{region_number}: the associated region number
  \item \code{regional_office}: the HHS regional office for the entire region
  \item \code{state_or_territory}: state or territory belonging to the region
}
}
\note{
Last updated 2015-08-09.
}
\references{
\url{http://www.hhs.gov/iea/regional/}
}
\keyword{datasets}

