% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmwr-map.r
\name{mmwr_week}
\alias{mmwr_week}
\title{Convert a Date to an MMWR day+week+year}
\usage{
mmwr_week(x)
}
\arguments{
\item{x}{a vector of \code{Date} objects or a character vector in \code{YYYY-mm-dd} format.}
}
\value{
data frame (tibble)
}
\description{
This is a reformat and re-export of a function in the \code{MMWRweek} package.
It provides a snake case version of its counterpart, produces a \code{tibble}
}
\examples{
mwk <- mmwr_week(Sys.Date())
}
