% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmwr-map.r
\name{mmwr_week_to_date}
\alias{mmwr_week_to_date}
\title{Convert an MMWR year+week or year+week+day to a Date object}
\usage{
mmwr_week_to_date(year, week, day = NULL)
}
\arguments{
\item{year, week, day}{Year, week and month vectors. All must be the same length
unless \code{day} is \code{NULL}.}
}
\value{
vector of \code{Date} objects
}
\description{
This is a reformat and re-export of a function in the \code{MMWRweek} package.
It provides a snake case version of its counterpart and produces a vector
of \code{Date} objects that corresponds to the input MMWR year+week or year+week+day
vectors. This also adds some parameter checking and cleanup to avoid exceptions.
}
\examples{
mwd <- mmwr_week_to_date(2016,10,3)
}
