% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmwr-map.r
\docType{data}
\name{mmwrid_map}
\alias{mmwrid_map}
\title{MMWR ID to Calendar Mappings}
\format{
A data frame with 4,592 rows and 4 columns
}
\description{
The CDC uses a unique "Morbidity and Mortality Weekly Report" identifier
for each week that starts at 1 (Ref: < https://www.cdc.gov/mmwr/preview/mmwrhtml/su6004a9.htm>).
This data frame consists of 4 columns:
\itemize{
\item \code{wk_start}: Start date (Sunday) for the week (\code{Date})
\item \code{wk_end}: End date (Saturday) for the week (\code{Date})
\item \code{year_wk_num}: The week of the calendar year
\item \code{mmwrid}: The unique MMWR identifier
These can be "left-joined" to data provided from the CDC to perform MMWR identifier
to date mappings.
}
}
