% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ili-weekly-state.r
\name{state_data_providers}
\alias{state_data_providers}
\title{Retrieve metadata about U.S. State CDC Provider Data}
\usage{
state_data_providers()
}
\description{
Retrieve metadata about U.S. State CDC Provider Data
}
\examples{
sdp <- state_data_providers()
}
