% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cdcfluview-package.r
\docType{package}
\name{cdcfluview}
\alias{cdcfluview}
\alias{cdcfluview-package}
\title{Retrieve Flu Season Data from the CDC FluView Portal}
\description{
The U.S. Centers for Disease Control (CDC) maintains a portal
\code{http://gis.cdc.gov/grasp/fluview/fluportaldashboard.html} for
accessing state, regional and national influenza statistics. The Flash
interface makes it difficult and time-consuming to select and retrieve
influenza data. This package provides functions to access the data provided
by portal's underlying API.
}
\author{
Bob Rudis (@hrbrmstr)
}

