% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/cdcfluview.R
\name{get_state_data}
\alias{get_state_data}
\title{Retrieves the state-level data from the CDC's FluView Portal}
\usage{
get_state_data(years = 2014)
}
\arguments{
\item{years}{a vector of years to retrieve data for (i.e. \code{2014} for CDC flu seasn 2014-2015)}
}
\value{
A \code{data.frame} of state-level data for the specified seasons
}
\description{
Uses the data source from the CDC' State-levelFluView \url{http://gis.cdc.gov/grasp/fluview/main.html}
and provides state flu reporting data as a single data frame
}
\examples{
\dontrun{
get_state_dat(2014)
get_state_data(c(2013,2014))
}
}

