% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospital.r
\name{hospitalizations}
\alias{hospitalizations}
\title{Laboratory-Confirmed Influenza Hospitalizations}
\usage{
hospitalizations(surveillance_area = c("flusurv", "eip", "ihsp"))
}
\arguments{
\item{surveillance_area}{one of "`flusurv`", "`eip`", or "`ihsp`"}
}
\description{
Laboratory-Confirmed Influenza Hospitalizations
}
\examples{
hosp_fs <- hospitalizations("flusurv")
hosp_eip <- hospitalizations("eip")
hosp_ihsp <- hospitalizations("ihsp")
}
\references{
- [Hospital Portal](https://gis.cdc.gov/GRASP/Fluview/FluHospRates.html)
}
